move into ftw_tools folder
