--- src/ftw/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ["datamodules", "datasets", "metrics", "models", "trainers", "utils"]
