--- src/ftw_cli/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ["cli", "download", "model", "unpack", "inference"]
