--- src/tests/data-files/min_config.yaml
@@ -1,45 +1,0 @@
-trainer:
-  max_epochs: 1
-  log_every_n_steps: 5
-  accelerator: cpu
-  default_root_dir: logs/FTW-CI
-  devices: 1
-  strategy: ddp_find_unused_parameters_true
-  callbacks:
-    - class_path: lightning.pytorch.callbacks.ModelCheckpoint
-      init_args:
-        monitor: val_loss
-        mode: min
-        save_top_k: 0
-        save_last: true
-        filename: "{epoch}-{val_loss:.2f}"
-model:
-  class_path: ftw.trainers.CustomSemanticSegmentationTask
-  init_args:
-    class_weights: [0.31,0.68]
-    loss: "ce" 
-    model: "unet"
-    backbone: "efficientnet-b3"
-    weights: true
-    patch_weights : false
-    in_channels: 8
-    num_classes: 2
-    num_filters: 64
-    ignore_index: 3
-    lr: 1e-3
-    patience: 100
-data:
-  class_path: ftw.datamodules.FTWDataModule
-  init_args:
-    batch_size: 32
-    num_workers: 4
-    train_countries:
-      - rwanda
-    val_countries:
-      - rwanda
-    test_countries:
-      - rwanda
-  dict_kwargs:
-    root: data/ftw
-    load_boundaries: false
-seed_everything: 7
